requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
python-barcode[images]==0.15.1
Pillow==11.1.0
pypdf==5.1.0
reportlab==4.2.5
python-docx==1.1.2
gunicorn==23.0.0
Werkzeug==3.1.3
